enforce 100/y citations cutoff
The relevance cutoff of each software requiring at least one year with
100 citations was introduced a posteriori, without removing historical
entries from the list that did not fulfil these criteria.

User feedback shows that this leads to confusion and can be viewed as
unfair (in particular by developers who would like to see their code
listed) an.
We agree, and hereby enforce the relevance cutoff also on historical
codes, removing those that do not pass it. Should any of these codes
pass the relevance threshold in the future, the git history should make
it easy to reintroduce them.