re-add DFTB+
Turns out that the query string "DFTB+" does *not* turn up all articles
containing the term "DFTB+".
In particular, the numbers of articles citing the DFTB+ articles
strongly exceeds the number of search results for the term "DFTB+".

Switching to citations of the DFTB+ article.

Note: No other codes in the list currently use a query string containing
the "+" character.