update query string for Gaussian
The original query string

  Gaussian package Frisch

often matched Gaussian in contexts other than the name of the software,
and 'package' in the name of other software.
I'm replacing this by a list of gaussian versions (which is how one
should cite Gaussian). As it turns out, the number of results even increases.