remove ASE/ASAP
ASAP is a standalone code hosted at https://wiki.fysik.dtu.dk/asap,
usually cited by providing this URL (and with <10 citations in 2020).

ASE is a python framework that can be used to run calculations with ASAP
but also with many others and does not qualify as an atomistic
simulation engine by itself.
The atomistic.software list may be extended to include workflow managers
like ASE going forward. In that case, however, we will make sure to
include all relevant ones.