remove citation data for 2007-2009 again
This was originally intended as a way of figuring out when each code
"entered the scene" (i.e. surpassed 100 citations per year).
However, due to the overall citation growth this may be a somewhat
problematic measure; also extending the time window back to the 2000s
requires a careful check in order to make sure we are not missing codes
that were relevant in that time frame.

We may get back to this at some later stage - the data is preserved in
git in case we want to reuse it then.