add benchmarks column
Following feedback by reviewers, add a "benchmarks" column.
In order not to expand the maintenance effort, to begin with this column
only registers whether a given benchmark exists (starting with the
delta-value from [1]).

Including actual values of benchmarks in the list could be envisioned
in cases where benchmark web sites offer APIs.

[1] https://molmod.ugent.be/deltacodesdft