underline abbreviations
Still using tooltip on hover but the abbreviations are now underlined
(and underline vanishes on hover).

Hover in general has the disadvantage of not supporting touch interfaces
but the alternative of displaying information on click requires to
transform all of the text in to buttons, which need to be listened to
etc.
It may work fine but it requires significant styling efforts.