Refresh citation data for 2020
Most of the citation data for the year 2020 was collected on April 10th
2021.
The estimate of total citation counts for codes with larger numbers of
citations all seemed to increase more strongly than during the preceding
years.

After refreshing the citation data for 2020 this apparent bias seems to
have disappeared.