split license field
The "license" field combined information about cost and source
availability, and did not allow for all combinations.
E.g. the ORCA code is free for academic use but the source code is not
available. This is not correctly recorded.

The split has further benefits, e.g. that the filter values are now
human-readable (the previous abbreviations were not explained anywhere).

Finally, the new "license" column now contains links to the license text
on SPDX whenever the license is an OSI-approved open-source license.